move useLogin to its own file. it earned it.
