+++ src/components/pages/Login.jsx
@@ -1,43 +1,8 @@
 import React, { useState, useEffect, useContext } from 'react'
 
+import { useLogin } from 'effects/useLogin'
 import { SessionContext } from 'providers/SessionProvider'
 import { Login as LoginOrganism } from 'components/organisms/Login'
-
-const url = 'https://api.poetnetwork.net/login'
-
-const fetchLogin = credentials => fetch(url, {
-  method: 'POST',
-  body: JSON.stringify(credentials),
-  headers: {
-    'content-type': 'application/json; charset=utf-8'
-  }
-})
-
-const useLogin = credentials => {
-  const [response, setResponse] = useState()
-  const [error, setError] = useState()
-  const [token, setToken] = useState()
-
-  useEffect(() => {
-    if (!credentials)
-      return
-    setError(null)
-    setToken(null)
-    fetchLogin(credentials).then(setResponse)
-  }, [credentials])
-
-  useEffect(() => {
-    if (response) {
-      if (response.status === 200) {
-        response.json().then(_ => _.token).then(setToken)
-      } else {
-        setError(response.statusText)
-      }
-    }
-  }, [response])
-
-  return { token, error }
-}
 
 export const Login = () => {
   const [credentials, setCredentials] = useState(null)
